+++ ClueNet.Core/Structures/SignalGroupManager.cs
@@ -1,12 +1,11 @@
-﻿using System;
+﻿using ClueNet.Core.Daq.Interfaces;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Text;
 
 namespace ClueNet.Core.Structures
 {
-    internal enum SignalState { None, Started, Completed }
-
     internal class SignalGroupManager
     {
         private ConcurrentDictionary<string, SignalGroup> SignalDict { get; set; }
@@ -14,12 +13,25 @@ namespace ClueNet.Core.Structures
 
         public void AddValue(string groupName, string signalName, double value)
         {
+            if (SignalDict.ContainsKey(groupName) == false)
+            {
+                SignalDict[groupName] = new SignalGroup();
+            }
+
             SignalDict[groupName].AddValue(signalName, value);
         }
 
         public SignalGroup this[string groupName]
         {
-            get { return SignalDict[groupName]; }
+            get 
+            {
+                if (SignalDict.ContainsKey(groupName) == false)
+                {
+                    SignalDict[groupName] = new SignalGroup();
+                }
+
+                return SignalDict[groupName]; 
+            }
         }
 
         public void Reset()
@@ -33,31 +45,22 @@ namespace ClueNet.Core.Structures
 
     internal class SignalGroup
     {
-        // 訊號名稱, 例如: Machine01
+        // 訊號群組名稱, 例如: Machine01
         public string Name { get; set; }
         private ConcurrentDictionary<string, SignalItem> SignalDict { get; set; } 
             = new ConcurrentDictionary<string, SignalItem>();
 
         // 是否接收數值
         private readonly object _lockOfReceiveEnabled = new object();
-        private bool _receiveEnabled;
-        public bool ReceiveEnabled
+        private SignalState _signalState;
+        public SignalState SignalState
         {
-            get { return _receiveEnabled; }
+            get { return _signalState; }
             set
             {
                 lock(_lockOfReceiveEnabled)
                 {
-                    if (value && (_receiveEnabled == false))
-                    {
-                        State = SignalState.Started;
-                    }
-                    else if ((value == false) && _receiveEnabled)
-                    {
-                        State = SignalState.Completed;
-                    }
-
-                    _receiveEnabled = value;
+                    _signalState = value;
                 }
             }
         }
@@ -78,7 +81,7 @@ namespace ClueNet.Core.Structures
 
         public void AddValue(string signalName, double value)
         {
-            if (ReceiveEnabled)
+            if (SignalState == SignalState.Start)
             {
                 if (SignalDict.ContainsKey(signalName) == false)
                 {
@@ -101,34 +104,44 @@ namespace ClueNet.Core.Structures
 
         public SignalItem this[string signalName]
         {
-            get { return SignalDict[signalName]; }
+            get
+            {
+                if (SignalDict.ContainsKey(signalName) == false)
+                {
+                    SignalDict[signalName] = new SignalItem();
+                }
+
+                return SignalDict[signalName]; 
+            }
         }
     }
 
     internal class SignalItem
     {
+        // 訊號單元名稱, 例如: 震動
         public string Name { get; set; }
         public string DisplayName { get; set; }
         private ConcurrentQueue<double> Values { get; set; } = new ConcurrentQueue<double>();
+        public int Count { get { return Values.Count; } }
 
         // 是否接收數值
         private readonly object _lockOfReceiveEnabled = new object();
-        private bool _receiveEnabled;
-        public bool ReceiveEnabled
+        private SignalState _signalState;
+        public SignalState SignalState
         {
-            get { return _receiveEnabled; }
+            get { return _signalState; }
             set
             {
                 lock (_lockOfReceiveEnabled)
                 {
-                    _receiveEnabled = value;
+                    _signalState = value;
                 }
             }
         }
 
         public void AddValue(double value)
         {
-            if (ReceiveEnabled)
+            if (SignalState == SignalState.Start)
             {
                 Values.Enqueue(value);
             }