+++ ClueNet.Tests/DeviceManagerTest.cs
@@ -7,24 +7,16 @@ namespace ClueNet.Tests
 {
     public class DeviceManagerTest : IDisposable
     {
-        private readonly DeviceManager _deviceManager;
+        private readonly DeviceManager _manager;
 
         public DeviceManagerTest()
         {
-            _deviceManager = new DeviceManager();
+            _manager = new DeviceManager();
         }
 
         public void Dispose()
         {
-            _deviceManager.Dispose();
-        }
-
-        [Fact]
-        public void CheckHealth_WhenDisconnected_ReturnFalse()
-        {
-            bool result = _deviceManager.CheckHealth();
-
-            Assert.False(result, "1 should not be prime");
+            _manager.Dispose();
         }
 
         [Fact]
@@ -33,8 +25,8 @@ namespace ClueNet.Tests
             // Arrange
             double actual = 0;
             double expected = 25;
-            _deviceManager.Initial();
-            _deviceManager.DataReceived += (object sender, DaqDataEventArgs e) =>
+            _manager.Initial();
+            _manager.DataReceived += (object sender, DaqDataEventArgs e) =>
             {
                 if (e.Name == "DemoTemperature")    // only TemperatureDemo
                 {
@@ -47,7 +39,7 @@ namespace ClueNet.Tests
             };
 
             // Act
-            _deviceManager.Connect();
+            _manager.Connect();
             System.Threading.Thread.Sleep(1200);
 
             // Assert
@@ -58,23 +50,24 @@ namespace ClueNet.Tests
         public void DiDevice_WhenConnected_IsCorrect()
         {
             // Arrange
-            bool actual = false;
-            bool expected = DateTime.Now.Second / 30 == 0;
-            _deviceManager.Initial();
-            _deviceManager.DigitalInputReceived += (object sender, DaqDigitalInputEventArgs e) =>
+            SignalState actual = SignalState.Complete;
+            SignalState expected = DateTime.Now.Second / 30 == 0 ? 
+                SignalState.Start : SignalState.Complete;
+            _manager.Initial();
+            _manager.DigitalInputReceived += (object sender, DaqControlEventArgs e) =>
             {
                 if (e.Name == "DemoTemperature")    // only TemperatureDemo
                 {
-                    actual = e.Enabled;
+                    actual = e.State;
                 }
                 else
                 {
-                    bool temp = e.Enabled;
+                    SignalState temp = e.State;
                 }
             };
 
             // Act
-            _deviceManager.Connect();
+            _manager.Connect();
             System.Threading.Thread.Sleep(1200);
 
             // Assert